+++ cmake/FindWEX.cmake
@@ -169,6 +169,7 @@ set(wex_LIBRARIES
   wex-stc${USE_DEBUG}
   wex-vi${USE_DEBUG}
   wex-ex${USE_DEBUG}
+  wex-ctags${USE_DEBUG}
   wex-ui${USE_DEBUG}
   wex-common${USE_DEBUG}
   wex-data${USE_DEBUG}