+++ include/wex/ctags/wex.h
@@ -1,5 +1,5 @@
 ////////////////////////////////////////////////////////////////////////////////
-// Name:      ex/wex.h
+// Name:      ctags/wex.h
 // Purpose:   General wex include file
 // Author:    Anton van Wezenbeek
 // Copyright: (c) 2022 Anton van Wezenbeek
@@ -7,11 +7,5 @@
 
 #pragma once
 
-#include <wex/ex/address.h>
-#include <wex/ex/addressrange.h>
-#include <wex/ex/command-parser.h>
-#include <wex/ex/ex-stream.h>
-#include <wex/ex/ex.h>
-#include <wex/ex/macro-mode.h>
-#include <wex/ex/macros.h>
-#include <wex/ex/variable.h>
+#include <wex/ctags/ctags-entry.h>
+#include <wex/ctags/ctags.h>