+++ include/wex/wex.h
@@ -9,12 +9,13 @@
 
 #include <wex/common/wex.h>
 #include <wex/core/wex.h>
+#include <wex/ctags/wex.h>
 #include <wex/data/wex.h>
 #include <wex/del/wex.h>
 #include <wex/ex/wex.h>
 #include <wex/factory/wex.h>
-#include <wex/syntax/wex.h>
 #include <wex/stc/wex.h>
+#include <wex/syntax/wex.h>
 #include <wex/ui/wex.h>
 #include <wex/vcs/wex.h>
 #include <wex/vi/wex.h>