+++ test/ctags/test-ctags.cpp
@@ -2,28 +2,20 @@
 // Name:      test-ctags.cpp
 // Purpose:   Implementation for wex unit testing
 // Author:    Anton van Wezenbeek
-// Copyright: (c) 2021 Anton van Wezenbeek
+// Copyright: (c) 2021-2022 Anton van Wezenbeek
 ////////////////////////////////////////////////////////////////////////////////
 
-#include <wex/ex/ctags.h>
-#include <wex/ex/ex.h>
+#include <wex/ctags/ctags.h>
 
 #include "test.h"
 
 TEST_CASE("wex::ctags")
 {
-  wex::data::stc data;
-
-  auto* ex = new wex::ex(get_stc());
-
-  SUBCASE("default")
+  SUBCASE("constructor")
   {
-    // default find
-    REQUIRE(wex::ctags::find("test_app"));
-    REQUIRE(!wex::ctags::find("xest_app"));
-
     // default auto_complete
-    REQUIRE(wex::ctags(ex).auto_complete("test_").starts_with("test_app"));
+    REQUIRE(
+      wex::ctags(get_stc()).auto_complete("test_").starts_with("test_app"));
 
     // setup a filter using find
     wex::ctags_entry filter;
@@ -32,24 +24,15 @@ TEST_CASE("wex::ctags")
     REQUIRE(filter.kind() == "f");
 
     // auto_complete using filter should now return member functions
-    REQUIRE(wex::ctags(ex).auto_complete("me", filter).starts_with("method"));
-    REQUIRE(wex::ctags(ex).auto_complete("he", filter).empty());
+    REQUIRE(
+      wex::ctags(get_stc()).auto_complete("me", filter).starts_with("method"));
+    REQUIRE(wex::ctags(get_stc()).auto_complete("he", filter).empty());
+
+    REQUIRE(wex::ctags(get_stc()).separator() != ' ');
   }
 
-  SUBCASE("non_existing_file")
+  SUBCASE("existing-ctags")
   {
-    REQUIRE(wex::ctags::close());
-    REQUIRE(!wex::ctags::close());
-
-    wex::ctags::open("xxx");
-
-    REQUIRE(!wex::ctags::find("test_app"));
-  }
-
-  SUBCASE("own_file")
-  {
-    REQUIRE(wex::ctags::close());
-
     wex::ctags::open("test-ctags");
 
     REQUIRE(!wex::ctags::find(""));
@@ -58,6 +41,16 @@ TEST_CASE("wex::ctags")
     REQUIRE(!wex::ctags::find("xxxx"));
     REQUIRE(wex::ctags::find("test_app"));
     REQUIRE(!wex::ctags::next());
-    REQUIRE(wex::ctags(ex).separator() != ' ');
+
+    REQUIRE(wex::ctags::close());
+    REQUIRE(!wex::ctags::close());
+  }
+
+  SUBCASE("non-existing-ctags")
+  {
+    wex::ctags::open("xxx");
+
+    REQUIRE(!wex::ctags::close());
+    REQUIRE(!wex::ctags::find("test_app"));
   }
 }